+++ theories/friendship_theorem.v
@@ -5,11 +5,11 @@ Unset Strict Implicit.
 Unset Printing Implicit Defensive.
 
 Require Import Friendship.combinatorics.
+Require Import Friendship.statement_reduction.
 
-(*
-Left: try change 'Co' from a total function to e.g.
-  Co: (u != v) -> {w | F u w & F v w} *)
-Theorem Friendship
+
+(* Longer version *)
+Theorem Friendship'
   (T: finType) (T_nonempty: [set: T] != set0)
   (F (* friendship relation *): rel T) (Fsym: symmetric F) (Firr: irreflexive F)
   (Co: T -> T -> T) (Col: forall u v : T, u != v -> F u (Co u v))
@@ -24,3 +24,18 @@ Proof.
 Qed.
 
   
+(* Shorter version *)
+Theorem Friendship
+  (T: finType) (T_nonempty: [set: T] != set0)
+  (F (* friendship relation *): rel T) (Fsym: symmetric F) (Firr: irreflexive F) :
+
+  (* u!= v have exactly 1 common friend *)
+  (forall (u v: T), u != v -> #|[set w | F u w & F v w]| == 1) ->
+  {u : T | forall v : T, u != v -> F u v}.
+Proof.
+  move => co_set.
+  apply: Friendship'; try by [].
+  exact (Col T_nonempty co_set).
+  exact (Cor T_nonempty co_set).
+  exact (CoUnique T_nonempty co_set).
+Qed.