+++ CHANGELOG.md
@@ -5,5 +5,6 @@
 - This changelog
 
 ### Changed
+- Made a shorter formulation without the required total functions; added `statement_reduction.v`.
 - Simplified `matrix_casts.v` after reading the castmx proofs in mathcomp.matrix. Turns out matrix casts are very simple in ssreflect.
 - sigW in `combinatorics` instead of my much more complicated proof.